nepal_weather/requirements.txt
Django>=4.2.0
requests>=2.28.0
python-decouple>=3.8
django-cors-headers>=4.0.0
Pillow>=9.0.0  